+++ main.rb
@@ -1,4 +1,5 @@
 require './app'
+require 'pry'
 
 def main
   app = App.new
